"Completed" exercise 6
really I didn't understand how to get things centered vertically
and horizontally on the height/width of the body. I could easily get
things figured out by utilizing more div containers but when it came
to centering the whole page I was quite lost because for example
when I use the body selector in css to create a flexbox, all the items
get put on one line which is not desireable.